server/requirements.txt
Django==4.0.4
django-extensions==3.1.5
django-ninja==0.17.0
django-cors-headers==3.12.0
